pyproject.toml
[project]
name = "miifachi-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp",
    "aiosqlite",
    "cogwatch>=3.3.1",
    "py-cord>=2.6.1",
    "python-dotenv",
]
